mock de infra layer

+++ lib/app/module/bank_statement/domain/entities/bank_transactions.dart
@@ -5,7 +5,7 @@ class BankTransactions extends Equatable {
   String? description;
   String? identification;
   String? date;
-  int? value;
+  double? value;
   double? balanceAfterOperation;
   int? operation;
